Finish splitting.
