+++ src/estimagic/logging/read_log.py
@@ -15,12 +15,12 @@ import numpy as np
 import pandas as pd
 from pybaum import tree_flatten, tree_unflatten
 
-from estimagic.logging.database_utilities import (
+from estimagic.logging.load_database import load_database
+from estimagic.logging.read_from_database import (
     read_last_rows,
     read_new_rows,
     read_specific_row,
 )
-from estimagic.logging.load_database import load_database
 from estimagic.parameters.tree_registry import get_registry
